Title: 사각형 채우기 3, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240630/사각형 채우기 3/rectangle-fill-3.cpp
@@ -13,7 +13,7 @@ int main() {
     int n;
     cin >> n;
 
-    vector<int> dp = vector<int>(MAX_N, 0);
+    vector<long long> dp = vector<long long>(MAX_N, 0);
 
     dp[0] = 1;
     dp[1] = 2;
